Getting the persistence tests to compile and run
